% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_to_omop.R
\name{etl_to_omop}
\alias{etl_to_omop}
\title{Synthea to OMOP ETL}
\usage{
etl_to_omop(
  user = NULL,
  password = NULL,
  server = NULL,
  port = NULL,
  omop_cdm_schema = "omop_cdm",
  synthea_schema = "native_synthea",
  path_to_synthea_csvs = "synthea_output/2020-12-20 20:12:56/csv/",
  path_to_omop_vocab_csvs = ""
)
}
\arguments{
\item{omop_cdm_schema}{Destination schema for the OMOP Vocabulary Tables and ETL'd Synthea data}

\item{synthea_schema}{Schema where the native synthea csvs will be loaded for the ETL.}

\item{path_to_synthea_csvs}{Path to the csvs produced by \code{\link{run_synthea}}.}

\item{path_to_omop_vocab_csvs}{Path to csvs downloaded the reconstituted (if CPT4 is desired) from \url{athena.ohdsi.org}.}

\item{connDetails}{Connection Details object, Default: cd}
}
\description{
Synthea to OMOP ETL
}
\seealso{
\code{\link[secretary]{typewrite}}
\code{\link[ETLSyntheaBuilder]{DropVocabTables}},\code{\link[ETLSyntheaBuilder]{DropEventTables}},\code{\link[ETLSyntheaBuilder]{DropSyntheaTables}},\code{\link[ETLSyntheaBuilder]{DropMapAndRollupTables}},\code{\link[ETLSyntheaBuilder]{CreateVocabTables}},\code{\link[ETLSyntheaBuilder]{CreateEventTables}},\code{\link[ETLSyntheaBuilder]{CreateSyntheaTables}},\code{\link[ETLSyntheaBuilder]{LoadSyntheaTables}},\code{\link[ETLSyntheaBuilder]{LoadVocabFromCsv}},\code{\link[ETLSyntheaBuilder]{CreateVocabMapTables}},\code{\link[ETLSyntheaBuilder]{CreateVisitRollupTables}},\code{\link[ETLSyntheaBuilder]{LoadEventTables}}
}
