% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_synthea.R
\name{run_synthea}
\alias{run_synthea}
\title{Run Synthea}
\usage{
run_synthea(
  seed,
  populationSize,
  moduleFilter,
  state,
  city,
  edit_synthea.properties = TRUE
)
}
\arguments{
\item{seed}{(Optional) Seed.}

\item{populationSize}{(Optional) Integer of simulated population size.}

\item{moduleFilter}{(Optional) Filter for a module. For example, "metabolic*" for all metabolic modules.}

\item{state}{(Optional) Population state. Defaults to Massachusetts.}

\item{city}{(Optional) Population city. If provided, \code{state} must be provided.}

\item{edit_synthea.properties}{If TRUE and the session is interactive, the synthea.properties file is opened for editing and the run is skipped. To run Synthea using the new properties, the function should be called again with this argument set to FALSE.}
}
\value{
Simulated data in a timestamped folder in a \code{synthea_output} path.
}
\description{
Run synthea according to the README guide found \href{here}{https://github.com/synthetichealth/synthea/blob/master/README.md}. The default setting opens the synthea.properties file to set the output parameters, such as setting the exporter.csv.export option to TRUE, which is required for \href{Synthea to OMOP builder}{https://github.com/OHDSI/ETL-Synthea}, which works off of the csv files and not the fhir files. To complete the run, this function should be re-called after saving any changes to the synthea.properties file and setting edit_synthea.properties to FALSE.
}
\seealso{
\code{\link[cli]{cat_line}}
\code{\link[glitter]{clone}}
\code{\link[secretary]{typewrite}}
}
